remove arvnetwork.h, use arvnetworkprivate.h only
