--- src/arvnetwork.h
@@ -1,56 +1,0 @@
-/* Aravis - Digital camera library
- *
- * Copyright © 2009-2019 Emmanuel Pacaud
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General
- * Public License along with this library; if not, write to the
- * Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
- * Boston, MA 02110-1301, USA.
- *
- * Author: Emmanuel Pacaud <emmanuel@gnome.org>
- */
-
-#ifndef ARV_NETWORK_H
-#define ARV_NETWORK_H
-
-#if !defined (ARV_H_INSIDE) && !defined (ARAVIS_COMPILATION)
-#error "Only <arv.h> can be included directly."
-#endif
-
-#include <arvtypes.h>
-
-// for socklen_t
-#include <gio/gnetworking.h>
-
-G_BEGIN_DECLS
-
-#define _ARV_IFACES
-
-#ifdef _ARV_IFACES
-GList* arv_enumerate_network_interfaces(void);
-#endif
-
-typedef struct _ArvNetworkInterface ArvNetworkInterface;
-
-struct sockaddr* arv_network_interface_get_addr(ArvNetworkInterface* a);
-struct sockaddr* arv_network_interface_get_netmask(ArvNetworkInterface* a);
-struct sockaddr* arv_network_interface_get_broadaddr(ArvNetworkInterface* a);
-const char* arv_network_interface_get_name(ArvNetworkInterface* a);
-void arv_network_interface_free(ArvNetworkInterface* a);
-
-#ifdef G_OS_WIN32
-// prototype is missing on mingw, let's reimplement
-const char *inet_ntop(int af, const void *src, char *dst, socklen_t cnt);
-#endif
-
-#endif
